requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# Requirements — Humanoid Robotic Head  (NIT Jamshedpur)
# Install: pip3 install -r requirements.txt --break-system-packages
# ─────────────────────────────────────────────────────────────────────────────

# ── AI / Cloud ────────────────────────────────────────────────────────────────
google-genai>=1.0.0

# ── Speech Recognition ────────────────────────────────────────────────────────
SpeechRecognition>=3.10.0
pyaudio>=0.2.14

# ── Text-to-Speech ────────────────────────────────────────────────────────────
gTTS>=2.5.0

# ── Audio Playback ────────────────────────────────────────────────────────────
pygame>=2.5.0

# ── PCA9685 / I2C Servo Driver ────────────────────────────────────────────────
adafruit-circuitpython-servokit>=1.3.0
adafruit-circuitpython-pca9685>=3.4.0
adafruit-blinka>=8.0.0
smbus2>=0.4.0

# ── Computer Vision ───────────────────────────────────────────────────────────
opencv-python>=4.8.0
ultralytics>=8.0.0

# ── Utilities ─────────────────────────────────────────────────────────────────
Pillow>=10.0.0
